create util file for calc props and import into slug
